% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{complexbrowser_changing_table_data}
\alias{complexbrowser_changing_table_data}
\title{Build top changing-complex table data}
\usage{
complexbrowser_changing_table_data(
  scored_database,
  condition,
  noise_threshold,
  fc_threshold
)
}
\arguments{
\item{scored_database}{Scored complex database.}

\item{condition}{Condition number.}

\item{noise_threshold}{Maximum allowed noise.}

\item{fc_threshold}{Fold-change threshold.}
}
\value{
A data frame for top up/down complexes.
}
\description{
Build top changing-complex table data
}
