% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{complexbrowser_complex_summary}
\alias{complexbrowser_complex_summary}
\title{Summarize scored complexes for a condition}
\usage{
complexbrowser_complex_summary(
  scored_database,
  no_cond,
  no_rep,
  condition,
  noise_threshold
)
}
\arguments{
\item{scored_database}{Scored complex database.}

\item{no_cond}{Number of experimental conditions.}

\item{no_rep}{Number of replicates per condition.}

\item{condition}{Condition number.}

\item{noise_threshold}{Maximum allowed noise.}
}
\value{
A character scalar summary.
}
\description{
Summarize scored complexes for a condition
}
