% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{complexbrowser_complex_table_filename}
\alias{complexbrowser_complex_table_filename}
\title{Build the complex-table CSV filename}
\usage{
complexbrowser_complex_table_filename(time = Sys.time())
}
\arguments{
\item{time}{Timestamp to include in the filename.}
}
\value{
A filename matching the Shiny download handler.
}
\description{
Build the complex-table CSV filename
}
