% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{complexbrowser_input_statistics_filename}
\alias{complexbrowser_input_statistics_filename}
\title{Build the input-statistics CSV filename}
\usage{
complexbrowser_input_statistics_filename(time = Sys.time())
}
\arguments{
\item{time}{Timestamp to include in the filename.}
}
\value{
A filename matching the Shiny download handler.
}
\description{
Build the input-statistics CSV filename
}
