% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{complexbrowser_load_complex_portal}
\alias{complexbrowser_load_complex_portal}
\title{Load the bundled prepared EBI Complex Portal database}
\usage{
complexbrowser_load_complex_portal()
}
\value{
A data frame containing prepared EBI Complex Portal complex records.
}
\description{
Load the bundled prepared EBI Complex Portal database
}
