% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{complexbrowser_load_corum}
\alias{complexbrowser_load_corum}
\title{Load the bundled prepared CORUM database}
\usage{
complexbrowser_load_corum()
}
\value{
A data frame containing prepared CORUM complex records.
}
\description{
Load the bundled prepared CORUM database
}
