% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{complexbrowser_prepare_qc_report_data}
\alias{complexbrowser_prepare_qc_report_data}
\title{Prepare QC report summary data}
\usage{
complexbrowser_prepare_qc_report_data(stats, no_cond, no_rep)
}
\arguments{
\item{stats}{QC statistics list.}

\item{no_cond}{Number of experimental conditions.}

\item{no_rep}{Number of replicates per condition.}
}
\value{
A list matching the legacy `QCreport.rmd` parameter data.
}
\description{
Prepare QC report summary data
}
