% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_analysis.R
\name{complexbrowser_prepare_user_database}
\alias{complexbrowser_prepare_user_database}
\title{Prepare a user-defined complex database}
\usage{
complexbrowser_prepare_user_database(csv_file)
}
\arguments{
\item{csv_file}{Path to a CSV file with complex database columns.}
}
\value{
A prepared database data frame.
}
\description{
Prepare a user-defined complex database
}
