% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{complexbrowser_run_app}
\alias{complexbrowser_run_app}
\title{Launch the packaged ComplexBrowser Shiny app}
\usage{
complexbrowser_run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to `shiny::runApp()`.}
}
\value{
The result of `shiny::runApp()`, invisibly.
}
\description{
Start the installed ComplexBrowser app from the package `inst/shiny`
directory.
}
\examples{
\dontrun{
complexbrowser_run_app()
}

}
