% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{complexbrowser_write_input_statistics}
\alias{complexbrowser_write_input_statistics}
\title{Write the input-statistics table as CSV}
\usage{
complexbrowser_write_input_statistics(table, file)
}
\arguments{
\item{table}{Input statistics table.}

\item{file}{Output CSV path.}
}
\value{
`file`, invisibly.
}
\description{
Write the input-statistics table as CSV
}
